requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
chromadb
langchain_groq
langchain_chroma